+++ .rubocop_todo.yml
@@ -158,7 +158,7 @@ Naming/MethodParameterName:
   Exclude:
     - 'app/helpers/spree/admin/base_helper_decorator.rb'
     - 'app/helpers/spree/base_helper_decorator.rb'
-    - 'app/services/subscription_validator.rb'
+    - 'engines/order_management/app/services/order_management/subscriptions/validator.rb'
     - 'lib/open_food_network/reports/bulk_coop_report.rb'
     - 'lib/open_food_network/xero_invoices_report.rb'
     - 'spec/lib/open_food_network/reports/report_spec.rb'
@@ -849,11 +849,6 @@ Style/FrozenStringLiteralComment:
     - 'app/services/reset_order_service.rb'
     - 'app/services/restart_checkout.rb'
     - 'app/services/search_orders.rb'
-    - 'app/services/subscription_estimator.rb'
-    - 'app/services/subscription_form.rb'
-    - 'app/services/subscription_validator.rb'
-    - 'app/services/subscription_variants_service.rb'
-    - 'app/services/subscriptions_count.rb'
     - 'app/services/tax_rate_finder.rb'
     - 'app/services/upload_sanitizer.rb'
     - 'app/services/variant_deleter.rb'
@@ -1350,11 +1345,6 @@ Style/FrozenStringLiteralComment:
     - 'spec/services/reset_order_service_spec.rb'
     - 'spec/services/restart_checkout_spec.rb'
     - 'spec/services/search_orders_spec.rb'
-    - 'spec/services/subscription_estimator_spec.rb'
-    - 'spec/services/subscription_form_spec.rb'
-    - 'spec/services/subscription_validator_spec.rb'
-    - 'spec/services/subscription_variants_service_spec.rb'
-    - 'spec/services/subscriptions_count_spec.rb'
     - 'spec/services/tax_rate_finder_spec.rb'
     - 'spec/services/upload_sanitizer_spec.rb'
     - 'spec/services/variants_stock_levels_spec.rb'
